move db files to folder
