--- CaloriesApplication/Database/Class2.cs
@@ -1,6 +1,0 @@
-﻿namespace CaloriesApplication.Database
-{
-    class Class2
-    {
-    }
-}
