+++ CaloriesApplication/Database/Product.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace CaloriesApplication
+namespace CaloriesApplication.Database
 {
 
     public class Product
